Add texture target and texture buffer access
