+++ crates/modor_graphics/src/lib.rs
@@ -35,11 +35,12 @@ pub use color::*;
 pub use material::default_2d::*;
 pub use material::*;
 pub use model::*;
-pub use resources::*;
 pub use runner::*;
+pub use shader::glob::*;
 pub use shader::*;
 pub use size::*;
 pub use target::*;
+pub use texture::glob::*;
 pub use texture::*;
 pub use window::*;
