+++ examples/examples/texture_target.rs
@@ -1,0 +1,6 @@
+#![allow(missing_docs)]
+
+#[modor::main]
+fn main() {
+    modor_examples::texture_target::main();
+}
