+++ examples/src/lib.rs
@@ -9,3 +9,4 @@
 
 pub mod custom_shader;
 pub mod rendering_2d;
+pub mod texture_target;
